+++ CHANGELOG.md
@@ -9,6 +9,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- `bin/r` executable for running Rails command inside the extension
+- `bin/sandbox_rails` executable for running Rails command inside the sandbox
 - Made Git ignore `sandbox` in generated extensions
 
 ## [1.0.1] - 2020-02-17